Enable feature and rename all v2 files in drugs
