+++ app/src/main/AndroidManifest.xml
@@ -38,11 +38,6 @@
       android:windowSoftInputMode="adjustResize" />
 
     <activity
-      android:name=".drugs.selectionv2.entry.CustomPrescriptionEntrySheetv2"
-      android:theme="@style/Clinic.V2.Theme.BottomSheetActivity"
-      android:windowSoftInputMode="adjustResize" />
-
-    <activity
       android:name=".bp.entry.BloodPressureEntrySheet"
       android:theme="@style/Clinic.V2.Theme.BottomSheetActivity"
       android:windowSoftInputMode="adjustResize" />
@@ -63,7 +58,7 @@
       android:windowSoftInputMode="adjustResize" />
 
     <activity
-      android:name=".drugs.selectionv2.dosage.DosagePickerSheet"
+      android:name=".drugs.selection.dosage.DosagePickerSheet"
       android:theme="@style/Clinic.V2.Theme.BottomSheetActivity"
       android:windowSoftInputMode="adjustResize" />
