+++ app/src/main/java/org/simple/clinic/activity/TheActivityComponent.kt
@@ -10,13 +10,10 @@ import dagger.Subcomponent
 import io.reactivex.Observable
 import org.simple.clinic.bp.entry.BloodPressureEntrySheet
 import org.simple.clinic.bp.entry.ConfirmRemoveBloodPressureDialog
-import org.simple.clinic.drugs.selection.ConfirmDeletePrescriptionDialog
-import org.simple.clinic.drugs.selection.PrescribedDrugsScreen
+import org.simple.clinic.drugs.selection.PrescribedDrugScreen
+import org.simple.clinic.drugs.selection.dosage.DosagePickerSheet
 import org.simple.clinic.drugs.selection.entry.CustomPrescriptionEntrySheet
-import org.simple.clinic.drugs.selectionv2.PrescribedDrugScreenV2
-import org.simple.clinic.drugs.selectionv2.dosage.DosagePickerSheet
-import org.simple.clinic.drugs.selectionv2.entry.CustomPrescriptionEntrySheetv2
-import org.simple.clinic.drugs.selectionv2.entry.confirmremovedialog.ConfirmRemovePrescriptionDialog
+import org.simple.clinic.drugs.selection.entry.confirmremovedialog.ConfirmRemovePrescriptionDialog
 import org.simple.clinic.editpatient.ConfirmDiscardChangesDialog
 import org.simple.clinic.editpatient.PatientEditScreen
 import org.simple.clinic.enterotp.EnterOtpScreen
@@ -71,11 +68,8 @@ interface TheActivityComponent {
   fun inject(target: PatientSearchScreen)
   fun inject(target: PatientSearchResultsScreen)
   fun inject(target: PatientSummaryScreen)
-  fun inject(target: PrescribedDrugsScreen)
   fun inject(target: BloodPressureEntrySheet)
   fun inject(target: CustomPrescriptionEntrySheet)
-  fun inject(target: CustomPrescriptionEntrySheetv2)
-  fun inject(target: ConfirmDeletePrescriptionDialog)
   fun inject(target: RegistrationPhoneScreen)
   fun inject(target: RegistrationFullNameScreen)
   fun inject(target: RegistrationPinScreen)
@@ -97,7 +91,7 @@ interface TheActivityComponent {
   fun inject(target: UpdatePhoneNumberDialog)
   fun inject(target: ConfirmRemoveBloodPressureDialog)
   fun inject(target: DosagePickerSheet)
-  fun inject(target: PrescribedDrugScreenV2)
+  fun inject(target: PrescribedDrugScreen)
   fun inject(target: ConfirmRemovePrescriptionDialog)
 
   @Subcomponent.Builder
