+++ app/src/main/java/org/simple/clinic/drugs/selection/dosage/DosageAdapter.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.drugs.selectionv2.dosage
+package org.simple.clinic.drugs.selection.dosage
 
 import android.view.LayoutInflater
 import android.view.View
