+++ app/src/main/java/org/simple/clinic/drugs/selection/entry/CustomPrescriptionEntryEvents.kt
@@ -1,9 +1,8 @@
 package org.simple.clinic.drugs.selection.entry
 
 import org.simple.clinic.widgets.UiEvent
-import java.util.UUID
 
-data class CustomPrescriptionSheetCreated(val patientUuid: UUID) : UiEvent {
+data class CustomPrescriptionSheetCreated(val openAs: OpenAs) : UiEvent {
   override val analyticsName = "Drugs:Custom:Show Prescription Sheet"
 }
 
@@ -19,6 +18,10 @@ data class CustomPrescriptionDrugDosageFocusChanged(val hasFocus: Boolean) : UiE
   override val analyticsName = "Drugs:Custom:Focused On Text Field"
 }
 
-class SaveCustomPrescriptionClicked : UiEvent {
+object SaveCustomPrescriptionClicked : UiEvent {
   override val analyticsName = "Drugs:Custom:Saved"
+}
+
+object RemoveCustomPrescriptionClicked : UiEvent {
+  override val analyticsName = "Drugs:Custom:Remove Clicked"
 }
