+++ app/src/main/java/org/simple/clinic/drugs/selection/entry/CustomPrescriptionEntrySheet.kt
@@ -3,9 +3,12 @@ package org.simple.clinic.drugs.selection.entry
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
+import android.view.View.GONE
+import android.view.View.VISIBLE
 import android.view.inputmethod.EditorInfo
 import android.widget.Button
-import android.widget.EditText
+import android.widget.TextView
+import com.google.android.material.textfield.TextInputEditText
 import com.jakewharton.rxbinding2.view.RxView
 import com.jakewharton.rxbinding2.widget.RxTextView
 import io.reactivex.Observable
@@ -15,16 +18,22 @@ import io.reactivex.subjects.PublishSubject
 import kotterknife.bindView
 import org.simple.clinic.R
 import org.simple.clinic.activity.TheActivity
+import org.simple.clinic.drugs.selection.entry.confirmremovedialog.ConfirmRemovePrescriptionDialog
 import org.simple.clinic.widgets.BottomSheetActivity
 import org.simple.clinic.widgets.UiEvent
+import org.simple.clinic.widgets.setTextAndCursor
+import org.simple.clinic.widgets.textChanges
 import java.util.UUID
 import javax.inject.Inject
 
 class CustomPrescriptionEntrySheet : BottomSheetActivity() {
 
-  private val drugNameEditText by bindView<EditText>(R.id.customprescription_drug_name)
-  private val drugDosageEditText by bindView<EditText>(R.id.customprescription_drug_dosage)
+  private val drugNameEditText by bindView<TextInputEditText>(R.id.customprescription_drug_name)
+  private val drugDosageEditText by bindView<TextInputEditText>(R.id.customprescription_drug_dosage)
   private val saveButton by bindView<Button>(R.id.customprescription_save)
+  private val enterMedicineTextView by bindView<TextView>(R.id.customprescription_enter_prescription)
+  private val editMedicineTextView by bindView<TextView>(R.id.customprescription_edit_prescription)
+  private val removeMedicineButton by bindView<Button>(R.id.customprescription_remove_button)
 
   @Inject
   lateinit var controller: CustomPrescriptionEntryController
@@ -42,7 +51,8 @@ class CustomPrescriptionEntrySheet : BottomSheetActivity() {
             drugNameChanges(),
             drugDosageChanges(),
             drugDosageFocusChanges(),
-            saveClicks())
+            saveClicks(),
+            removeClicks())
         .observeOn(io())
         .compose(controller)
         .observeOn(mainThread())
@@ -56,8 +66,8 @@ class CustomPrescriptionEntrySheet : BottomSheetActivity() {
   }
 
   override fun onBackgroundClick() {
-    val drugNameEmpty = drugNameEditText.text.isEmpty()
-    val dosageEmpty = drugDosageEditText.text.isEmpty()
+    val drugNameEmpty = drugNameEditText.text.isNullOrEmpty()
+    val dosageEmpty = drugDosageEditText.text.isNullOrEmpty()
         || drugDosageEditText.text.toString().trim() == getString(R.string.customprescription_dosage_placeholder)
 
     if (drugNameEmpty && dosageEmpty) {
@@ -66,17 +76,13 @@ class CustomPrescriptionEntrySheet : BottomSheetActivity() {
   }
 
   private fun sheetCreates(): Observable<UiEvent> {
-    val patientUuid = intent.getSerializableExtra(KEY_PATIENT_UUID) as UUID
-    return Observable.just(CustomPrescriptionSheetCreated(patientUuid))
+    val openAs = intent.getParcelableExtra(KEY_OPEN_AS) as OpenAs
+    return Observable.just(CustomPrescriptionSheetCreated(openAs))
   }
 
-  private fun drugNameChanges() = RxTextView.textChanges(drugNameEditText)
-      .map(CharSequence::toString)
-      .map(::CustomPrescriptionDrugNameTextChanged)
+  private fun drugNameChanges() = drugNameEditText.textChanges(::CustomPrescriptionDrugNameTextChanged)
 
-  private fun drugDosageChanges() = RxTextView.textChanges(drugDosageEditText)
-      .map(CharSequence::toString)
-      .map(::CustomPrescriptionDrugDosageTextChanged)
+  private fun drugDosageChanges() = drugDosageEditText.textChanges(::CustomPrescriptionDrugDosageTextChanged)
 
   private fun drugDosageFocusChanges() = RxView.focusChanges(drugDosageEditText)
       .map(::CustomPrescriptionDrugDosageFocusChanged)
@@ -86,8 +92,13 @@ class CustomPrescriptionEntrySheet : BottomSheetActivity() {
 
     return RxView.clicks(saveButton)
         .mergeWith(dosageImeClicks)
-        .map { SaveCustomPrescriptionClicked() }
+        .map { SaveCustomPrescriptionClicked }
   }
+
+  private fun removeClicks(): Observable<UiEvent> =
+      RxView
+          .clicks(removeMedicineButton)
+          .map { RemoveCustomPrescriptionClicked }
 
   fun setSaveButtonEnabled(enabled: Boolean) {
     saveButton.isEnabled = enabled
@@ -102,12 +113,46 @@ class CustomPrescriptionEntrySheet : BottomSheetActivity() {
     drugDosageEditText.post { drugDosageEditText.setSelection(0) }
   }
 
-  companion object {
-    private const val KEY_PATIENT_UUID = "patientUuid"
+  fun showEnterNewPrescriptionTitle() {
+    enterMedicineTextView.visibility = VISIBLE
+  }
 
-    fun intent(context: Context, patientUuid: UUID): Intent {
+  fun showEditPrescriptionTitle() {
+    editMedicineTextView.visibility = VISIBLE
+  }
+
+  fun showRemoveButton() {
+    removeMedicineButton.visibility = VISIBLE
+  }
+
+  fun hideRemoveButton() {
+    removeMedicineButton.visibility = GONE
+  }
+
+  fun setMedicineName(drugName: String) {
+    drugNameEditText.setTextAndCursor(drugName)
+  }
+
+  fun setDosage(dosage: String?) {
+    drugDosageEditText.setTextAndCursor(dosage ?: "")
+  }
+
+  fun showConfirmRemoveMedicineDialog(prescribedDrugUuid: UUID) {
+    ConfirmRemovePrescriptionDialog.showForPrescription(prescribedDrugUuid, supportFragmentManager)
+  }
+
+  companion object {
+    private const val KEY_OPEN_AS = "openAs"
+
+    fun intentForAddNewPrescription(context: Context, patientUuid: UUID): Intent {
       val intent = Intent(context, CustomPrescriptionEntrySheet::class.java)
-      intent.putExtra(KEY_PATIENT_UUID, patientUuid)
+      intent.putExtra(KEY_OPEN_AS, OpenAs.New(patientUuid))
+      return intent
+    }
+
+    fun intentForUpdatingPrescription(context: Context, prescribedDrugUuid: UUID): Intent {
+      val intent = Intent(context, CustomPrescriptionEntrySheet::class.java)
+      intent.putExtra(KEY_OPEN_AS, OpenAs.Update(prescribedDrugUuid))
       return intent
     }
   }
