+++ app/src/main/java/org/simple/clinic/summary/PatientSummaryConfig.kt
@@ -3,6 +3,5 @@ package org.simple.clinic.summary
 data class PatientSummaryConfig(
     val numberOfBpPlaceholders: Int,
     val numberOfBpsToDisplay: Int,
-    val isUpdatePhoneDialogEnabled: Boolean,
-    val isNewPrescriptionScreenEnabled: Boolean
+    val isUpdatePhoneDialogEnabled: Boolean
 )
