+++ app/src/main/java/org/simple/clinic/summary/PatientSummaryScreenController.kt
@@ -265,12 +265,9 @@ class PatientSummaryScreenController @Inject constructor(
 
     return events
         .ofType<PatientSummaryUpdateDrugsClicked>()
-        .withLatestFrom(patientUuid, configProvider.toObservable())
-        .map { (_, patientUuid, config) ->
-          when (config.isNewPrescriptionScreenEnabled) {
-            true -> { ui: Ui -> ui.showUpdatePrescribedDrugsScreenv2(patientUuid) }
-            false -> { ui: Ui -> ui.showUpdatePrescribedDrugsScreenv1(patientUuid) }
-          }
+        .withLatestFrom(patientUuid)
+        .map { (_, patientUuid) ->
+          { ui: Ui -> ui.showUpdatePrescribedDrugsScreen(patientUuid) }
         }
   }
