--- app/src/main/res/drawable/ic_delete_20dp.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-  android:width="20dp"
-  android:height="20dp"
-  android:tint="#FFFFFF"
-  android:viewportHeight="24.0"
-  android:viewportWidth="24.0">
-  <path
-    android:fillColor="#FF000000"
-    android:pathData="M6,19c0,1.1 0.9,2 2,2h8c1.1,0 2,-0.9 2,-2V7H6v12zM19,4h-3.5l-1,-1h-5l-1,1H5v2h14V4z" />
-</vector>
