--- app/src/main/res/drawable/selector_prescribeddrug_background.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-
-  <item android:state_checked="true">
-    <inset android:insetBottom="12dp" android:insetTop="12dp">
-      <shape android:shape="rectangle">
-        <solid android:color="?colorAccent" />
-        <corners android:radius="14dp" />
-      </shape>
-    </inset>
-  </item>
-
-  <item>
-    <inset android:insetBottom="12dp" android:insetTop="12dp">
-      <shape android:shape="rectangle">
-        <solid android:color="#ECEFF0" />
-        <corners android:radius="14dp" />
-      </shape>
-    </inset>
-  </item>
-</selector>
