+++ app/src/main/res/layout/list_prescribeddrugs_custom_drug.xml
@@ -1,52 +1,43 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.cardview.widget.CardView android:id="@+id/prescribeddrug_item_customdrug_rootlayout"
+  xmlns:android="http://schemas.android.com/apk/res/android"
+  xmlns:app="http://schemas.android.com/apk/res-auto"
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="wrap_content"
-  android:background="@color/white"
-  android:orientation="vertical">
+  android:background="@color/white100"
+  android:clickable="true"
+  android:focusable="true"
+  android:foreground="?selectableItemBackground"
+  android:orientation="vertical"
+  app:cardCornerRadius="0dp">
 
   <LinearLayout
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
+    android:layout_gravity="center_vertical"
     android:orientation="horizontal"
-    android:paddingBottom="@dimen/spacing_12"
-    android:paddingEnd="@dimen/spacing_16"
-    android:paddingStart="@dimen/spacing_16"
-    android:paddingTop="@dimen/spacing_12">
+    android:padding="@dimen/spacing_16">
 
     <TextView
       android:id="@+id/prescribeddrug_item_customdrug_name"
+      style="@style/Clinic.V2.TextAppearance.Body1Left.Grey0"
       android:layout_width="0dp"
       android:layout_height="wrap_content"
-      android:layout_gravity="center_vertical"
       android:layout_marginEnd="@dimen/spacing_8"
       android:layout_weight="1"
-      android:textSize="@dimen/textsize_16"
-      tools:text="Reese's" />
+      tools:text="Amlodipine" />
 
     <TextView
       android:id="@+id/prescribeddrug_item_customdrug_dosage"
-      android:layout_width="0dp"
-      android:layout_height="wrap_content"
-      android:layout_gravity="center_vertical"
-      android:layout_weight="0.8"
-      android:textAllCaps="false"
-      android:textAppearance="@style/TextAppearance.AppCompat.Widget.Button"
-      tools:text="10 packets" />
-
-    <ImageButton
-      android:id="@+id/prescribeddrug_item_customdrug_delete"
+      style="@style/Clinic.V2.TextAppearance.Body1Left.Blue1"
       android:layout_width="wrap_content"
       android:layout_height="wrap_content"
-      android:background="?attr/selectableItemBackgroundBorderless"
-      android:contentDescription="@string/prescribeddrug_delete_custom_drug"
-      android:padding="@dimen/spacing_4"
-      android:src="@drawable/ic_delete_20dp"
-      android:tint="@color/gray_500" />
+      tools:text="10 mg" />
   </LinearLayout>
 
   <View
-    style="@style/Clinic.Separator.VerticalContent"
+    android:id="@+id/protocoldrug_item_divider"
+    style="@style/Clinic.V2.ListSeparator"
     android:layout_gravity="bottom" />
 </androidx.cardview.widget.CardView>
