+++ app/src/main/res/layout/sheet_custom_prescription_entry.xml
@@ -1,49 +1,101 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+  xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="wrap_content"
   android:gravity="center_horizontal"
   android:orientation="vertical">
 
-  <TextView
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:layout_marginBottom="@dimen/spacing_16"
-    android:layout_marginTop="@dimen/spacing_16"
-    android:text="@string/customprescription_enter_medicine"
-    android:textAllCaps="true"
-    android:textColor="@color/black_opacity_75"
-    android:textSize="14sp" />
+  <RelativeLayout
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content">
+
+    <TextView
+      android:id="@+id/customprescription_enter_prescription"
+      style="@style/Clinic.V2.TextAppearance.H6.Grey0"
+      android:layout_width="wrap_content"
+      android:layout_height="wrap_content"
+      android:layout_centerHorizontal="true"
+      android:layout_marginBottom="@dimen/spacing_16"
+      android:layout_marginTop="@dimen/spacing_24"
+      android:text="@string/customprescription_enter_medicine"
+      android:visibility="gone" />
+
+    <TextView
+      android:id="@+id/customprescription_edit_prescription"
+      android:layout_width="wrap_content"
+      android:layout_height="wrap_content"
+      android:layout_alignParentStart="true"
+      android:layout_centerVertical="true"
+      android:layout_marginBottom="@dimen/spacing_16"
+      android:layout_marginStart="@dimen/spacing_16"
+      android:layout_marginTop="@dimen/spacing_24"
+      android:layout_toStartOf="@id/customprescription_remove_button"
+      android:ellipsize="end"
+      android:lines="1"
+      android:text="@string/customprescription_edit_medicine"
+      android:textAppearance="@style/Clinic.V2.TextAppearance.H6.Grey0"
+      android:visibility="gone"
+      tools:ignore="UnusedAttribute"
+      tools:visibility="visible" />
+
+    <Button
+      android:id="@+id/customprescription_remove_button"
+      style="@style/Widget.AppCompat.Button.Borderless"
+      android:layout_width="wrap_content"
+      android:layout_height="wrap_content"
+      android:layout_alignParentEnd="true"
+      android:layout_centerVertical="true"
+      android:layout_marginBottom="@dimen/spacing_16"
+      android:layout_marginEnd="@dimen/spacing_16"
+      android:layout_marginTop="@dimen/spacing_24"
+      android:text="@string/bloodpressureentry_remove"
+      android:textAppearance="@style/Clinic.V2.TextAppearance.Button2.Red1"
+      android:visibility="gone"
+      tools:visibility="visible" />
+
+  </RelativeLayout>
 
   <LinearLayout
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:layout_marginBottom="@dimen/spacing_24"
     android:layout_marginEnd="@dimen/spacing_24"
-    android:layout_marginStart="@dimen/spacing_24"
+    android:layout_marginStart="@dimen/spacing_16"
+    android:baselineAligned="true"
     android:orientation="horizontal">
 
-    <EditText
-      android:id="@+id/customprescription_drug_name"
+    <com.google.android.material.textfield.TextInputLayout
       android:layout_width="0dp"
       android:layout_height="wrap_content"
-      android:layout_weight="0.6"
-      android:hint="@string/customprescription_drug_name_hint"
-      android:imeOptions="actionNext"
-      android:inputType="textCapWords"
-      android:textSize="20sp" />
+      android:layout_weight="0.6">
 
-    <EditText
-      android:id="@+id/customprescription_drug_dosage"
+      <com.google.android.material.textfield.TextInputEditText
+        android:id="@+id/customprescription_drug_name"
+        style="@style/Clinic.V2.TextAppearance.Body1Left.Grey0"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:hint="@string/customprescription_drug_name_hint"
+        android:imeOptions="actionNext"
+        android:inputType="textCapWords" />
+    </com.google.android.material.textfield.TextInputLayout>
+
+    <com.google.android.material.textfield.TextInputLayout
       android:layout_width="0dp"
       android:layout_height="wrap_content"
-      android:layout_marginStart="@dimen/spacing_16"
-      android:layout_weight="0.4"
-      android:hint="@string/customprescription_drug_dosage_hint"
-      android:imeOptions="actionDone"
-      android:inputType="textCapSentences"
-      android:text="@string/customprescription_dosage_placeholder"
-      android:textSize="20sp" />
+      android:layout_weight="0.4">
+
+      <com.google.android.material.textfield.TextInputEditText
+        android:id="@+id/customprescription_drug_dosage"
+        style="@style/Clinic.V2.TextAppearance.Body1Left.Grey0"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="@dimen/spacing_16"
+        android:hint="@string/customprescription_drug_dosage_hint"
+        android:imeOptions="actionDone"
+        android:inputType="textCapSentences"
+        android:text="@string/customprescription_dosage_placeholder" />
+    </com.google.android.material.textfield.TextInputLayout>
   </LinearLayout>
 
   <Button
