+++ app/src/main/res/values/colors.xml
@@ -11,9 +11,6 @@
 
   <color name="patientsummary_edit_label_color">@color/blue1</color>
 
-  <color name="prescribeddrugs_protocol_drug_selected">@color/white</color>
-  <color name="prescribeddrugs_protocol_drug_unselected">@color/color_accent</color>
-
   <color name="textinputlayout_error">@color/color_error</color>
   <color name="snackbar_action">@color/green2</color>
 </resources>
