+++ app/src/main/res/values/strings.xml
@@ -174,7 +174,6 @@
   <!-- Prescribed drugs -->
   <string name="patient_prescribeddrugs_screen_title">BP Medicines</string>
   <string name="patient_prescribeddrugs_add_custom_prescription">Add another medicine</string>
-  <string name="prescribeddrug_delete_custom_drug">Delete this prescription</string>
 
   <string name="customprescription_enter_medicine">Enter medicine</string>
   <string name="customprescription_edit_medicine">Edit medicine</string>
