+++ app/src/test/java/org/simple/clinic/drugs/selection/PrescribedDrugsScreenControllerTest.kt
@@ -7,11 +7,14 @@ import com.nhaarman.mockito_kotlin.whenever
 import io.reactivex.Completable
 import io.reactivex.Observable
 import io.reactivex.subjects.PublishSubject
+import junitparams.JUnitParamsRunner
+import junitparams.Parameters
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
+import org.junit.runner.RunWith
 import org.simple.clinic.drugs.PrescriptionRepository
-import org.simple.clinic.drugs.selection.ProtocolDrugSelectionListItem.DosageOption
+import org.simple.clinic.drugs.selection.entry.CustomPrescribedDrugListItem
 import org.simple.clinic.facility.FacilityRepository
 import org.simple.clinic.patient.PatientMocker
 import org.simple.clinic.protocol.ProtocolDrugAndDosages
@@ -22,12 +25,13 @@ import org.simple.clinic.util.RxErrorsRule
 import org.simple.clinic.widgets.UiEvent
 import java.util.UUID
 
+@RunWith(JUnitParamsRunner::class)
 class PrescribedDrugsScreenControllerTest {
 
   @get:Rule
   val rxErrorsRule = RxErrorsRule()
 
-  private val screen = mock<PrescribedDrugsScreen>()
+  private val screen = mock<PrescribedDrugScreen>()
   private val userSession = mock<UserSession>()
   private val facilityRepository = mock<FacilityRepository>()
   private val protocolRepository = mock<ProtocolRepository>()
@@ -85,61 +89,21 @@ class PrescribedDrugsScreenControllerTest {
     uiEvents.onNext(PrescribedDrugsScreenCreated(patientUuid))
 
     val expectedUiModels = listOf(
-        ProtocolDrugSelectionListItem(
+        ProtocolDrugListItem(
             id = 0,
             drugName = amlodipine10mg.name,
-            option1 = DosageOption.Unselected(amlodipine5mg),
-            option2 = DosageOption.Selected(amlodipine10mg, prescription = amlodipine10mgPrescription)),
-        ProtocolDrugSelectionListItem(
+            prescribedDrug = amlodipine10mgPrescription,
+            hideDivider = false),
+        ProtocolDrugListItem(
             id = 1,
             drugName = telmisartan40mg.name,
-            option1 = DosageOption.Unselected(telmisartan40mg),
-            option2 = DosageOption.Unselected(telmisartan80mg)),
+            prescribedDrug = null,
+            hideDivider = false),
         CustomPrescribedDrugListItem(telmisartan9000mgPrescription),
         CustomPrescribedDrugListItem(reesesPrescription),
         CustomPrescribedDrugListItem(fooPrescription),
         CustomPrescribedDrugListItem(barPrescription))
     verify(screen).populateDrugsList(expectedUiModels)
-  }
-
-  @Test
-  fun `when a protocol drug is selected then a prescription should be saved for it`() {
-    whenever(userSession.requireLoggedInUser()).thenReturn(Observable.never())
-    whenever(prescriptionRepository.savePrescription(any(), any())).thenReturn(Completable.complete())
-    whenever(protocolRepository.drugsForProtocolOrDefault(any())).thenReturn(Observable.never())
-    whenever(prescriptionRepository.newestPrescriptionsForPatient(patientUuid)).thenReturn(Observable.empty())
-
-    val amlodipine10mg = PatientMocker.protocolDrug(name = "Amlodipine", dosage = "10mg")
-
-    uiEvents.onNext(PrescribedDrugsScreenCreated(patientUuid))
-    uiEvents.onNext(ProtocolDrugDosageSelected(amlodipine10mg))
-
-    verify(prescriptionRepository).savePrescription(patientUuid, drug = amlodipine10mg)
-  }
-
-  @Test
-  fun `when a protocol drug is unselected then its prescription should be soft deleted`() {
-    val amlodipine5mg = PatientMocker.protocolDrug(name = "Amlodipine", dosage = "5mg")
-    val amlodipine10mg = PatientMocker.protocolDrug(name = "Amlodipine", dosage = "10mg")
-    val unselectedPrescriptionId = UUID.randomUUID()
-
-    whenever(userSession.requireLoggedInUser()).thenReturn(Observable.never())
-    whenever(prescriptionRepository.savePrescription(any(), any())).thenReturn(Completable.complete())
-    whenever(prescriptionRepository.softDeletePrescription(unselectedPrescriptionId)).thenReturn(Completable.complete())
-
-    whenever(protocolRepository.drugsForProtocolOrDefault(any())).thenReturn(Observable.never())
-    whenever(prescriptionRepository.newestPrescriptionsForPatient(patientUuid)).thenReturn(Observable.empty())
-
-    uiEvents.onNext(PrescribedDrugsScreenCreated(patientUuid))
-    uiEvents.onNext(ProtocolDrugDosageSelected(amlodipine10mg))
-    uiEvents.onNext(ProtocolDrugDosageSelected(amlodipine5mg))
-    uiEvents.onNext(ProtocolDrugDosageUnselected(
-        drug = amlodipine10mg,
-        prescription = PatientMocker.prescription(uuid = unselectedPrescriptionId, name = amlodipine10mg.name, dosage = amlodipine10mg.dosage)))
-
-    verify(prescriptionRepository).savePrescription(patientUuid, drug = amlodipine10mg)
-    verify(prescriptionRepository).savePrescription(patientUuid, drug = amlodipine5mg)
-    verify(prescriptionRepository).softDeletePrescription(unselectedPrescriptionId)
   }
 
   @Test
@@ -149,16 +113,34 @@ class PrescribedDrugsScreenControllerTest {
     whenever(prescriptionRepository.newestPrescriptionsForPatient(patientUuid)).thenReturn(Observable.empty())
 
     uiEvents.onNext(PrescribedDrugsScreenCreated(patientUuid))
-    uiEvents.onNext(AddNewPrescriptionClicked())
+    uiEvents.onNext(AddNewPrescriptionClicked)
 
     verify(screen).showNewPrescriptionEntrySheet(patientUuid)
   }
 
+  @Parameters(
+      "Amlodipine",
+      "Telimisartan",
+      "Athenlol"
+  )
   @Test
-  fun `when a custom prescription delete is clicked then delete confirmation dialog should be shown`() {
-    val prescription = PatientMocker.prescription(name = "Amlodipine", dosage = "10mg")
-    uiEvents.onNext(DeleteCustomPrescriptionClicked(prescription))
+  fun `when a protocol drug is selected then open dosages sheet for that drug`(drugName: String) {
+    whenever(userSession.requireLoggedInUser()).thenReturn(Observable.never())
+    whenever(prescriptionRepository.savePrescription(any(), any())).thenReturn(Completable.complete())
+    whenever(protocolRepository.drugsForProtocolOrDefault(any())).thenReturn(Observable.never())
+    whenever(prescriptionRepository.newestPrescriptionsForPatient(patientUuid)).thenReturn(Observable.empty())
 
-    verify(screen).showDeleteConfirmationDialog(prescription)
+    uiEvents.onNext(PrescribedDrugsScreenCreated(patientUuid))
+    uiEvents.onNext(ProtocolDrugClicked(drugName = drugName, prescriptionForProtocolDrug = null))
+
+    verify(screen).showDosageSelectionSheet(drugName = drugName, patientUuid = patientUuid, prescribedDrugUuid = null)
+  }
+
+  @Test
+  fun `when a custom prescription is clicked then open upate custom prescription screen`() {
+    val prescribedDrug = PatientMocker.prescription()
+    uiEvents.onNext(CustomPrescriptionClicked(prescribedDrug))
+
+    verify(screen).showUpdateCustomPrescriptionSheet(prescribedDrug)
   }
 }
