+++ app/src/test/java/org/simple/clinic/drugs/selection/entry/CustomPrescriptionEntryControllerTest.kt
@@ -3,10 +3,12 @@ package org.simple.clinic.drugs.selection.entry
 import com.nhaarman.mockito_kotlin.any
 import com.nhaarman.mockito_kotlin.eq
 import com.nhaarman.mockito_kotlin.mock
+import com.nhaarman.mockito_kotlin.never
 import com.nhaarman.mockito_kotlin.times
 import com.nhaarman.mockito_kotlin.verify
 import com.nhaarman.mockito_kotlin.whenever
 import io.reactivex.Completable
+import io.reactivex.Observable
 import io.reactivex.subjects.PublishSubject
 import junitparams.JUnitParamsRunner
 import junitparams.Parameters
@@ -15,6 +17,7 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.simple.clinic.drugs.PrescriptionRepository
+import org.simple.clinic.patient.PatientMocker
 import org.simple.clinic.util.RxErrorsRule
 import org.simple.clinic.util.nullIfBlank
 import org.simple.clinic.widgets.UiEvent
@@ -29,6 +32,7 @@ class CustomPrescriptionEntryControllerTest {
   private val sheet = mock<CustomPrescriptionEntrySheet>()
   private val prescriptionRepository = mock<PrescriptionRepository>()
   private val patientUuid = UUID.randomUUID()
+  private val prescriptionUuid = UUID.randomUUID()
 
   private val uiEvents = PublishSubject.create<UiEvent>()
   private lateinit var controller: CustomPrescriptionEntryController
@@ -54,16 +58,17 @@ class CustomPrescriptionEntryControllerTest {
 
   @Test
   @Parameters(value = ["", "10mg"])
-  fun `when save is clicked then a new prescription should be saved`(dosage: String) {
+  fun `when sheet is opened in new mode and save is clicked then a new prescription should be saved`(dosage: String) {
     whenever(prescriptionRepository.savePrescription(patientUuid, "Amlodipine", dosage.nullIfBlank(), rxNormCode = null, isProtocolDrug = false))
         .thenReturn(Completable.complete())
 
-    uiEvents.onNext(CustomPrescriptionSheetCreated(patientUuid))
+    uiEvents.onNext(CustomPrescriptionSheetCreated(OpenAs.New(patientUuid)))
     uiEvents.onNext(CustomPrescriptionDrugNameTextChanged("Amlodipine"))
     uiEvents.onNext(CustomPrescriptionDrugDosageTextChanged(dosage))
-    uiEvents.onNext(SaveCustomPrescriptionClicked())
+    uiEvents.onNext(SaveCustomPrescriptionClicked)
 
     verify(prescriptionRepository).savePrescription(patientUuid, "Amlodipine", dosage.nullIfBlank(), rxNormCode = null, isProtocolDrug = false)
+    verify(prescriptionRepository, never()).updatePrescription(any())
     verify(sheet).finish()
   }
 
@@ -92,5 +97,96 @@ class CustomPrescriptionEntryControllerTest {
     uiEvents.onNext(CustomPrescriptionDrugDosageFocusChanged(true))
 
     verify(sheet).moveDrugDosageCursorToBeginning()
+  }
+
+  @Test
+  @Parameters(method = "params for showing title")
+  fun `when sheet is created then correct title should be populated`(openAs: OpenAs, showNewEntryTitle: Boolean) {
+    whenever(prescriptionRepository.prescription(any())).thenReturn(Observable.never())
+
+    uiEvents.onNext(CustomPrescriptionSheetCreated(openAs))
+
+    if (showNewEntryTitle) {
+      verify(sheet).showEnterNewPrescriptionTitle()
+    } else {
+      verify(sheet).showEditPrescriptionTitle()
+    }
+  }
+
+  @Suppress("Unused")
+  private fun `params for showing title`(): List<List<Any>> {
+    return listOf(
+        listOf(OpenAs.New(patientUuid), true),
+        listOf(OpenAs.Update(prescriptionUuid), false)
+    )
+  }
+
+  @Parameters(method = "params for showing remove button")
+  @Test
+  fun `the remove button should when the sheet is opened for edit`(openAs: OpenAs, showRemoveButton: Boolean) {
+    whenever(prescriptionRepository.prescription(any())).thenReturn(Observable.never())
+
+    uiEvents.onNext(CustomPrescriptionSheetCreated(openAs))
+
+    if (showRemoveButton) {
+      verify(sheet).showRemoveButton()
+    } else {
+      verify(sheet).hideRemoveButton()
+    }
+  }
+
+  @Suppress("Unused")
+  private fun `params for showing remove button`(): List<List<Any>> {
+    return listOf(
+        listOf(OpenAs.New(patientUuid), false),
+        listOf(OpenAs.Update(prescriptionUuid), true)
+    )
+  }
+
+  @Test
+  fun `when sheet is opened to edit prescription then the drug name and dosage should be pre-filled`() {
+    val prescription = PatientMocker.prescription(uuid = prescriptionUuid)
+    whenever(prescriptionRepository.prescription(prescriptionUuid)).thenReturn(Observable.just(prescription))
+
+    uiEvents.onNext(CustomPrescriptionSheetCreated(OpenAs.Update(prescriptionUuid)))
+
+    verify(sheet).setMedicineName(prescription.name)
+    verify(sheet).setDosage(prescription.dosage)
+  }
+
+  @Test
+  fun `when sheet is opened in edit mode and save is clicked after making changes, then the prescription should be updated`() {
+    val prescribedDrug = PatientMocker.prescription(uuid = prescriptionUuid, name = "Atnlol", dosage = "20mg")
+
+    whenever(prescriptionRepository.prescription(prescriptionUuid)).thenReturn(Observable.just(prescribedDrug))
+    whenever(prescriptionRepository.updatePrescription(any())).thenReturn(Completable.complete())
+
+    uiEvents.onNext(CustomPrescriptionSheetCreated(OpenAs.Update(prescriptionUuid)))
+    uiEvents.onNext(CustomPrescriptionDrugNameTextChanged("Atenolol"))
+    uiEvents.onNext(CustomPrescriptionDrugDosageTextChanged("5mg"))
+    uiEvents.onNext(SaveCustomPrescriptionClicked)
+
+    verify(prescriptionRepository).updatePrescription(prescribedDrug.copy(name = "Atenolol", dosage = "5mg"))
+    verify(prescriptionRepository, never()).savePrescription(any(), any())
+    verify(sheet).finish()
+  }
+
+  @Test
+  fun `when remove is clicked, the prescription should be deleted`() {
+    whenever(prescriptionRepository.prescription(any())).thenReturn(Observable.never())
+
+    uiEvents.onNext(CustomPrescriptionSheetCreated(OpenAs.Update(prescriptionUuid)))
+    uiEvents.onNext(RemoveCustomPrescriptionClicked)
+
+    verify(sheet).showConfirmRemoveMedicineDialog(prescriptionUuid)
+  }
+
+  @Test
+  fun `when prescription is deleted then close the sheet`() {
+    whenever(prescriptionRepository.prescription(prescriptionUuid)).thenReturn(Observable.just(PatientMocker.prescription(uuid = prescriptionUuid, isDeleted = true)))
+
+    uiEvents.onNext(CustomPrescriptionSheetCreated(OpenAs.Update(prescriptionUuid)))
+
+    verify(sheet).finish()
   }
 }
