+++ app/src/test/java/org/simple/clinic/drugs/selection/entry/confirmremovedialog/ConfirmRemovePrescriptionDialogControllerTest.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.drugs.selectionv2.entry.confirmremovedialog
+package org.simple.clinic.drugs.selection.entry.confirmremovedialog
 
 import com.nhaarman.mockito_kotlin.mock
 import com.nhaarman.mockito_kotlin.verify
